Support only one endpoint
To support multiple endpoints use a DNS record that returns multiple IP addresses
